lookup.go: check realm name. For a unused case, this is a [security fix]
The security fix in this commit affects clients whose trusted verifiers
use the same signing key and monitor colluding malicious keyservers.

The issue does NOT affect clients whose verifiers use different signing
keys for different keyservers. I am not aware of any verifiers who reuse
signing keys, so hopefully nobody is affected.

coname.VerifyLookup used to ignore the realm name in signed epoch heads.
This means that a malicious keyserver may have been able to have
verifiers verify two copies of itself (with different contents and
different realm names) and then present whichever view it chooses.

The client configuration file now includes the canonical realm name for
each name and VerifyLookup considers all SignedEpochHeads that do not
match that name to be invalid.

I changed the protobuf field numbers in config.proto. This should not
matter because this file is only used as a JSON schema; no actual
protobufs adhering to this spec are used anywhere in this codebase.

This change should be ported to the JS implementation of VerifyLookup.